Reworked the whole concept again.
Shifted the focus on the inner workings of the `Lexicon` type.
Added two subtypes: `Entry` and `Page`.
Was able to remove many other external types.
Renamed some types, to have more precise and expressive names.
Still need to figure out how to replace the `Vocable` type in order to have a way to have a smart-type for the `Lexicon`'s `storage`.